QGtkOpenGLContext: Implement single-threaded OpenGL rendering
This is a slow but functional implementation of OpenGL rendering to GTK
windows. There are challenges in matching GDK's model here, but it could
be improved significantly from what this commit does at least.

For now, the approach is to render into independent contexts on the Qt
side, and glReadPixels/re-upload the framebuffer into GTK's context.
That is dumb, but it's necessary because of how context sharing is used
on the GTK side, and Qt's implicit API requirement of contexts being
independent of surfaces.

Multithreaded renderer support will be fixed in a following commit.