Remove GtkGLArea and draw buffer directly to cairo
GtkGLArea was entirely unnecessary here -- it would essentially end up
uploading our rendered and downloaded data to a texture, drawing it,
readng it back out, and raster-blitting it onto the cairo surface.

Skip that entire process by directly drawing our rendered GL content
onto the cairo surface, just like the non-GL windows do. Once GTK is
able to implement pure-OpenGL windows this should be improved, but for
now it's about as good as you can do in a GTK window.