QGtkSystemTrayIcon: Remove systray icon support
These have been deprecated for quite a while, and are increasingly
removed outright. Furthermore, they are a pretty bad UX anyway, so let's
do our part in not perpetrating it.

We return 'true' from systray available, so that messages can still be
delivered, until such a time as notifications have their own API
uncoupled from the systray support in QPA.