QGtkWindow: Fixes to keyboard handling
Go back to fetching text directly from Qt. Using the keyval is wrong in
many cases (e.g. modifier keys), and we cannot really fix all those
cases ourselves.

Additionally, alter the keyval when it's < 255. This seems to be
expected for e.g. shortcut activation, and thus fixes things like Ctrl+A
in a text widget.