QGtkOpenGLContext: Set device pixel ratio on the QImage
This wasn't needed before because we treated the QImage in the GL
context as a dumb buffer of bytes which was then read out in
QGtkWindow's draw handler, but with the advent of RasterGLSurface
support, the QImage is also shared with the QGtkBackingStore. So, if the
QGtkBackingStore creates the image, we must set the correct device pixel
ratio to ensure that it is rendered into correctly by QtWidgets.