QGtkWindow: Make window geometry thread-safe
This has to be done in a thread-safe manner, as it is accessed from
QGtkOpenGLContext to set up the FBO. It worked a lot of the time, but
exploded when resizing QtWebEngine examples for instance, so let's do
this properly.

Unfortunately, this exposes another bug: for some reason, gtk does not
send a configure-event for popup-style windows (which we use in menus).
To work around this, also connect to size-allocate to set a size on the
window.