Introduce QGtkEventDispatcher
This is a (potentially temporarily?) forked version of the glib QPA from
upstream. The upstream dispatcher has priority levels that cause some
examples (like qopenglwidget) to not render because Qt's events are
exhausting those from gtk.

For the time being, we copy it over and set our own priority levels. We
might also need a source for window system events specifically, for the
time being, processing them "as usual" seems to work OK.