QGtkWindow: Proper fix for priority issues
We hacked up a fix for setVisible, but this was the wrong fix. The
problem is that now everything is going through our event loop, and we
have set ourselves at a low priority, gtk is processing events before us
in many cases, leading to an initial frame of flicker.

To fix this, increase the priority of QWindowSystemEvent dispatching,
and while we're at it, improve the priority of other sources too.