QGtkWindow: improve geometry change handling
Send an expose event if the window size changes (and flush, so we draw
up-to-date content).

Coincidentally, this removes the menu hack (as we're now getting the
content size always, not the window size, as geometry should be). This
also seems to fix a few strange examples (like ricochet's add contact
dialog's initial size).