QGtkMenu: Yes, yet another menu rework.
The old implementation was quite good, but had some mysterious bugs with
menus disappearing in some applications related to GtkWidget parenting.
So we take a lesson from the former approach, and get rid of the idea of
trying to cache widgets, and instead, just make sure we don't create
them too often by queueing calls to generate the menus.