QGtkWindow: Make sense of geometry and frameMargins
This fixes window geometry and frameMargins to correctly account for
server-side and client-side decorations as well as GTK decorations (like
menubars). Resolves some bugs around the position of popup windows, and
probably others we hadn't noticed.

It remains to be determined whether setGeometry is actually correct. I
suspect it's not.