QGtkScreen: Only divide by dpr when needed
In 30b8b6aec3a1b5ef6373e842238409346d620a10 we divided the returned
resolution by the display scale. This is necessary under Wayland due to
a gtk+ bug, but doing this of course breaks x11.

A bug has been filed for this upstream
(https://bugzilla.gnome.org/show_bug.cgi?id=788497), but in the
meantime, only do the division on Wayland to avoid breaking X11.