Revert "QGtkWindow: Display all popup-type windows as popups"
This reverts commit bdbf1acab48b0447c94c9336d5c44fe96db21ef2.

This reintroduces the problem of focus stealing, which is most certainly
not good, but it seems like some tool-type windows (in particular,
Telegram Desktop's media viewer) expect to be able to
steal focus, and this matches up with xcb's behaviour.

Unfortunately, I'm unable to check exactly why notification windows
steal focus, as upstream Telegram has significantly changed and the code
in question no longer exists (in favour of libnotify support).