gtkextras: Add first extra; QGtkHeaderBar
QGtkHeaderBar can be used to add a GtkHeaderBar widget to a window.
Right now, it's a dead simple wrapper that just retrieves the GtkWidget
for you to then use yourself. It might make sense to add some simple
conveniences on top, but on the other hand, perhaps it's better not to.